# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SCCC IT Department Reporting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List weekly reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, finalized]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create/finalize a weekly report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    patch:
      operationId: updateReport
      tags: [reports]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a weekly report (CIO only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /reports/{id}/tickets:
    get:
      operationId: listReportTickets
      tags: [reports]
      summary: Solved Zendesk tickets during this report's week
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportTicketsResponse"
  /reports/{id}/extras:
    get:
      operationId: getReportExtras
      tags: [reports]
      summary: Week-scoped extras (PIRs, maintenance, goal progress) for the report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportExtras"
  /reports/{id}/email:
    post:
      operationId: emailReport
      tags: [reports]
      summary: Email the report (PDF/DOCX attached) to recipients (CIO only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailReportBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailReportResponse"
        "503":
          description: Email not configured
  /reports/{id}/finalize:
    post:
      operationId: finalizeReport
      tags: [reports]
      summary: Finalize and aggregate a weekly report (CIO only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/aggregate:
    get:
      operationId: getAggregateReport
      tags: [reports]
      summary: Get aggregated report data for a week
      parameters:
        - name: weekOf
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AggregateReport"

  # RISKS AND ISSUES
...
components:
  schemas:
...
    Entry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userRole:
          type: string
        category:
          type: string
          enum: [helpdesk, network, security, general]
        title:
          type: string
        description:
          type: string
        accomplishments:
          type: string
        challenges:
          type: string
        supportNeeded:
          type: string
        ticketCount:
          type: integer
        weekOf:
          type: string
          description: ISO date string of week start
        entryDate:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        isSubmitted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        [
          id,
          userId,
          category,
          title,
          description,
          weekOf,
          isSubmitted,
          createdAt,
        ]
...
    Report:
      type: object
      properties:
        id:
          type: integer
        weekOf:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [draft, finalized]
        summary:
          type: string
        accomplishments:
          type: string
        challenges:
          type: string
        strategicProgress:
          type: string
        nextWeekPlans:
          type: string
        metrics:
          type: object
          additionalProperties: true
        selectedItemIds:
          type: array
          items:
            type: integer
        selectedAfterActionIds:
          type: array
          items:
            type: integer
          nullable: true
        selectedMaintenanceIds:
          type: array
          items:
            type: string
          nullable: true
        selectedRiskIds:
          type: array
          description: |
            Open risk IDs explicitly included in this report. null = include all open risks
            in the export (default behavior). Empty array = include none.
          items:
            type: integer
          nullable: true
        projectIds:
          type: array
          description: IDs of projects linked to this report.
          items:
            type: integer
        includeGoalProgress:
          type: boolean
        includeOpenRisks:
          type: boolean
        includeCloudInventory:
          type: boolean
        emailRecipients:
          type: array
          items:
            type: string
        lastEmailedAt:
          type: string
          format: date-time
          nullable: true
        contributorCount:
          type: integer
        entryCount:
          type: integer
        createdBy:
          type: integer
        createdByName:
          type: string
        finalizedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, weekOf, status, createdAt]
    CreateReportBody:
      type: object
      properties:
        weekOf:
          type: string
        title:
          type: string
        summary:
          type: string
        accomplishments:
          type: string
        challenges:
          type: string
        strategicProgress:
          type: string
        nextWeekPlans:
          type: string
      required: [weekOf]
    UpdateReportBody:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        accomplishments:
          type: string
        challenges:
          type: string
        strategicProgress:
          type: string
        nextWeekPlans:
          type: string
        status:
          type: string
          enum: [draft, finalized]
        selectedItemIds:
          type: array
          items:
            type: integer
          nullable: true
        selectedAfterActionIds:
          type: array
          items:
            type: integer
          nullable: true
        selectedMaintenanceIds:
          type: array
          items:
            type: string
          nullable: true
        selectedRiskIds:
          type: array
          items:
            type: integer
          nullable: true
        includeGoalProgress:
          type: boolean
        includeOpenRisks:
          type: boolean
        includeCloudInventory:
          type: boolean
        emailRecipients:
          type: array
          items:
            type: string
            format: email
    ReportExtras:
      type: object
      properties:
        weekOf:
          type: string
        weekStart:
          type: string
        weekEnd:
          type: string
        afterActionReports:
          type: array
          items:
            type: object
            additionalProperties: true
        maintenance:
          type: array
          items:
            type: object
            additionalProperties: true
        goalProgress:
          type: array
          items:
            type: object
            additionalProperties: true
        cloudInventory:
          $ref: "#/components/schemas/CloudInventorySnapshot"
      required: [weekOf, afterActionReports, maintenance, goalProgress]
    EmailReportBody:
      type: object
      properties:
        recipients:
          type: array
          items:
            type: string
            format: email
        subject:
          type: string
        message:
          type: string
        format:
          type: string
          enum: [pdf, docx]
      required: [recipients]
    EmailReportResponse:
      type: object
      properties:
        sent:
          type: boolean
        recipients:
          type: array
          items:
            type: string
        filename:
          type: string
    ReportTicket:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        status:
          type: string
        assigneeName:
          type: string
          nullable: true
        updatedAt:
          type: string
        url:
          type: string
      required: [id, subject, status, url]
    ReportTicketsResponse:
      type: object
      properties:
        weekOf:
          type: string
        count:
          type: integer
        configured:
          type: boolean
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/ReportTicket"
      required: [weekOf, count, tickets]
    AggregateReport:
      type: object
      properties:
        weekOf:
          type: string
        totalEntries:
          type: integer
        contributorCount:
          type: integer
        eligibleUserIds:
          type: array
          description: User IDs whose weekly logs satisfy the shared department-report submission policy.
          items:
            type: integer
        byRole:
          type: object
          additionalProperties:
            type: integer
        byCategory:
          type: object
          additionalProperties:
            type: integer
        totalTickets:
          type: integer
        openRisks:
          type: integer
        entrySummaries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
        risks:
          type: array
          items:
            $ref: "#/components/schemas/Risk"
      required: [weekOf, totalEntries, contributorCount, eligibleUserIds]
...
    AzureVmRiskSummary:
      type: object
      properties:
        publicIp:
          type: integer
        unhealthy:
          type: integer
        retiringSize:
          type: integer
        flaggedVms:
          type: integer
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
      required:
        [publicIp, unhealthy, retiringSize, flaggedVms, high, medium, low]
    CloudInventorySnapshot:
      type: object
      properties:
        configured:
          type: boolean
        total:
          type: integer
        byStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
        risk:
          $ref: "#/components/schemas/AzureVmRiskSummary"
        lastSync:
          nullable: true
          type: object
          properties:
            status:
              type: string
            createdAt:
              type: string
              format: date-time
            createdCount:
              type: integer
            updatedCount:
              type: integer
            removedCount:
              type: integer
            changedCount:
              type: integer
            totalCount:
              type: integer
            error:
              type: string
              nullable: true
            actorName:
              type: string
              nullable: true
      required: [configured, total, byStatus, risk, lastSync]
...